Create computed license state in license composable
